+++ src/utils/referenceType.ts
@@ -1,6 +1,10 @@
 import { showQuickPick } from './vscode'
 
-const REFERENCE_TYPES = ['Symbol', 'Filename'] as const
+const REFERENCE_TYPES = [
+  'Symbol',
+  'Filename (with line)',
+  'Filename (no line)',
+] as const
 const REFERENCE_TYPE_PLACEHOLDER = 'Select reference type'
 
 export type ReferenceType = (typeof REFERENCE_TYPES)[number]
@@ -17,6 +21,14 @@ function isReferenceType(type?: string): type is ReferenceType {
 export async function getReferenceType(
   configType: string | undefined
 ): Promise<ReferenceType | undefined> {
+  /**
+   * Handle backward compatibility for old "Filename" type; pre-v1.2.0.
+   * Default old "Filename" to "Filename (with line)" for backward compatibility
+   */
+  if (configType === 'Filename') {
+    return 'Filename (with line)'
+  }
+
   if (isReferenceType(configType)) {
     return configType
   } else {